+++ src/controllers/postsController.ts
@@ -1,13 +1,15 @@
 import type { Request, Response, NextFunction } from "express";
 import { Prisma } from "../db/generated/prisma/client.ts";
 import { prisma } from "../db/prisma.ts";
+import { validationResult, matchedData } from "express-validator";
 
 async function postPost(req: Request, res: Response) {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ error: errors.array() });
+  }
+  const body = matchedData(req);
   try {
-    const body = req.body;
-    if (!body.title) {
-      return res.status(400).json({ error: "No title found" });
-    }
     const newPost = await prisma.post.create({
       data: {
         title: body.title,
@@ -36,9 +38,13 @@ async function getPosts(req: Request, res: Response) {
     return res.status(500).json({ error });
   }
 }
+
 async function getPost(req: Request, res: Response) {
+  const id = Number(req.params.id);
+  if (!id) {
+    return res.status(400).json({ error: "Invalid parameter" });
+  }
   try {
-    const id = Number(req.params.id);
     const post = await prisma.post.findUnique({
       where: {
         id: id,
@@ -52,9 +58,13 @@ async function getPost(req: Request, res: Response) {
     return res.status(500).json({ error });
   }
 }
+
 async function deletePost(req: Request, res: Response, next: NextFunction) {
+  const id = Number(req.params.id);
+  if (!id) {
+    return res.status(400).json({ error: "Invalid parameter" });
+  }
   try {
-    const id = Number(req.params.id);
     const deletedPost = await prisma.post.delete({
       where: {
         id: id,
@@ -71,10 +81,18 @@ async function deletePost(req: Request, res: Response, next: NextFunction) {
     return res.status(500).json({ error });
   }
 }
+
 async function updatePost(req: Request, res: Response, next: NextFunction) {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ error: errors.array() });
+  }
+  const body = matchedData(req);
+  const id = Number(req.params.id);
+  if (!id) {
+    return res.status(400).json({ error: "Invalid parameter" });
+  }
   try {
-    const body = req.body;
-    const id = Number(req.params.id);
     const updatedPost = await prisma.post.update({
       where: {
         id: id,
